requirements.txt
python-dotenv==1.0.0
python-telegram-bot==20.7
PyPDF2==3.0.1
chromadb==1.0.12
openai==0.28.1
sentence-transformers==2.2.2
langchain-text-splitters==0.0.1
striprtf==0.0.26